+++ src/models/access.rs
@@ -2,9 +2,9 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct MemnixAccess {
-    pub id: i8,
-    pub user_id: i8,
-    pub deck_id: i8,
+    pub id: u32,
+    pub user_id: u32,
+    pub deck_id: u32,
     pub permission: i8,
     
 }
