+++ src/models/revision.rs
@@ -2,8 +2,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct MemnixRevision {
-    pub user_id: i8,
-    pub card_id: i8,
+    pub user_id: u32,
+    pub card_id: u32,
     pub result: bool,
     pub result_int: i8,
     pub quality: i8
