+++ src/models/user.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct MemnixUser {
-    pub id: i8,
+    pub id: u32,
     pub user_name: String,
     pub discord_id: String,
 }
